Initial project
Issue #1 - We have a new project, with two little sprit (a red square and a black square). There is also a TestScene and a few prefab (One for the floor and one for a character). I add two little script : one for the controls and another one with a class which represents a character in Unity.