Story #3 Add automatic generation of testScene
- Modified prefabs

The generated map is a square with 3 platforms
The player is automaticaly instanciated at a currently hardcoded spawn location
The collisions are handled with BoxCollider2D

An additional saveScene function has been added. It allows to store the
generated scene into a new file. It is then possible to reload the
specific scene into unity.